Move CuratedDataQueryBuilder into base image
